remove unused attrib
Removing ['sumologic']['sources']['default_timezone'] as it's blocking
the new definition where ['sumologic']['sources'] is an array instead
of a hash and maps directly to the sources used in the sumo.json.

As far as I can tell, this attribute was never, ever
used for anything in this cookbook.

And it's marked as required!!!