Add support for setting the sources list via attributes
This allows you to specify the complete source list for a given host
in env/role/node/cookbook attributes.

Note that until the client side of the REST API for updating a hosts
sources list is implemented in this cookbook, this will only allow
setting of the sources list at install time.