Update recipe files for Windows support
Updated the recipe files as necessary to support installing the Windows
collector.

Changes include:
- Replace file references to `/etc/sumo.*` to instead pull from the
value from the `sumo_*_path` attributes
- Added directory resources to create the parent directory for both the
`sumo.conf` and `sumo.json` files (on Linux this is /etc, which is safe
to assume exists.  On Windows we default to C:\sumo, which likely does
not exist by default).
- Pull the installer command from the `installerCmd` attribute instead
of hard coding in `install.rb`
- For file and directory resources, only define `owner`, `group`, and
`mode` values on Linux (otherwise WARNings are generated on Windows