Move sumologic_collector_installer to a more Cheffy sumologic_collector resource
The sumo_collector resource has an install_and_configure resource which does the same thing as the current sumologic_collector_installer does.
This resource is far more inline with the Chef way.

I have kept the existing sumologic_collector_installer resource for backwards compatibility (to be removed later), it is now just a shim for the new resource