:configure action for sumologic_collector resource
Pairs with the :install resource, configures an installed Collector.
Ideally the collector being configured won't have been registered yet as some of these settings only work on first
registration. A use case for this is AMI baking, the collector can be baked into an AMI with the :install action,
and then configured on instance launch using the :configure action.

For normal workflow there is the :install_and_configure action which is the default