Delay the service restart until the end of the Chef run
There seems to be a race condition in the collector where, if it's restarted
while in the middle of registering, it can end up registering twice. Delaying
the restart until the end of the Chef run should help with that, as well as
limiting the total number of restarts if resources in a wrapper cookbook
are sending restart notifications for other things as well.