Leave the service disabled after the collector resource's :install action
The installer script, even if it's run with `-VskipRegistration=true`, enables
the service. This means the collector resource's `:install` action leaves
behind an unconfigured collector that will nonetheless try (and fail) to start
at boot time.

I can see the argument that this is the install script's default behavior, so
this is the state it should be left in. But it was a small enough change I
wanted to submit it and see what folks think.