allow sources to use a full path with the name attribute
Users of the `zap` cookbook will find that the way we currently create file collector definitions that if you use `zap_directory` on `/etc/sumo.json` all resources will be deleted. The reason for this is that there are some reasonable assumptions made in the zap cookbook that a file, template, etc resource should be derived in the following manner:
1. If the resource responds to `.path?()` then it will use that
1. Else it assumes that its the resource name

I toyed with the idea of keying off of the name and passing it to `path` but given that this would be more brittle I decided to stick with the ability to be explicit.

Signed-off-by: Ben Abrams <me@benabrams.it>